pyproject.toml
[project]
name = "telegrambot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "gunicorn>=23.0.0",
    "pymongo[srv]>=4.15.3",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
